Use the new dedicated Synapse API
